requirements.txt
# Core Dependencies
vaderSentiment>=3.3.2
textblob>=0.17.1

# Optional: For advanced sentiment analysis (uncomment if needed)
# transformers>=4.30.0
# torch>=2.0.0

# UI and Formatting
colorama>=0.4.6
rich>=13.0.0

# Data and Configuration
pyyaml>=6.0

# Visualization (Optional but recommended)
matplotlib>=3.7.0

# Web Application
flask>=3.0.0
flask-cors>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking (optional)
# mypy>=1.5.0

